% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/approx.backscale.pts.R
\name{approx.backscale.points}
\alias{approx.backscale.points}
\title{approx.backscale.points}
\usage{
approx.backscale.points(mind, maxd, pts.to.rescale, bitdepth)
}
\arguments{
\item{XXXX}{}
}
\value{
XXXX
}
\description{
XXXX
}
\details{
Approximately backscale bit depth scales values to wiah they were or would have been in the original
64 bit physical scale. Handy for when using the 8-bit only openCV hole filling routene.
}
\examples{
XXXX
}
\references{
XXXX
}

