% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coord2mat.R
\name{coords2mat}
\alias{coords2mat}
\title{coords2mat}
\usage{
coords2mat(x.idxs, y.idxs, z.vec, x.max.idx = NULL, y.max.idx = NULL)
}
\arguments{
\item{XXXX}{}
}
\value{
XXXX
}
\description{
Convert a vector of z's (heights, intensities, etc.) to a matrix
}
\details{
XXXX
}
\examples{
XXXX
}
\references{
XXXX
}

