% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/crop.circular.roi.R
\name{crop.circular.roi}
\alias{crop.circular.roi}
\title{crop.circular.roi}
\usage{
crop.circular.roi(dmat, xc.idx, yc.idx, pix.radius, crop.empty.border = TRUE,
  roi.type = "inner")
}
\arguments{
\item{XXXX}{}
}
\value{
XXXX
}
\description{
XXXX
}
\details{
Fill holes with openCV findContour and inPaint functionality. Back transform 8-bit result to approximate
64-bit physical scale of original data.
}
\examples{
XXXX
}
\references{
XXXX
}

