% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/crop.rectangular.roi.R
\name{crop.retangular.roi}
\alias{crop.retangular.roi}
\title{crop.ractangular.roi}
\usage{
crop.retangular.roi(dmat, left.idx.start = NULL, right.idx.start = NULL,
  top.idx.start = NULL, bot.idx.start = NULL, clip.empty.border = TRUE,
  roi.type = "inner")
}
\arguments{
\item{XXXX}{}
}
\value{
XXXX
}
\description{
XXXX
}
\details{
XXXX
}
\examples{
XXXX
}
\references{
XXXX
}

