% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/display2d.openCV.R
\name{display2D.openCV}
\alias{display2D.openCV}
\title{display2D_openCV}
\usage{
display2D.openCV(dmat, num.rows, num.cols, bit.depth, file.ext,
  printQ = FALSE)
}
\arguments{
\item{XXXX}{}
}
\value{
XXXX
}
\description{
Wrapper for Display2D_openCV C++ function
}
\details{
Does some error checking and gets system file path to openCV executable in inst/display2d
}
\examples{
XXXX
}
\references{
XXXX
}

