% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fill.holes2.R
\name{fill.holes2}
\alias{fill.holes2}
\title{fill.holes2}
\usage{
fill.holes2(dmat, big.hole.tol.percent)
}
\arguments{
\item{XXXX}{}
}
\value{
XXXX
}
\description{
XXXX
}
\details{
Fill holes with openCV findContour and inPaint functionality. Back transform 8-bit result to approximate
64-bit physical scale of original data.
}
\examples{
XXXX
}
\references{
XXXX
}

