% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FindContours}
\alias{FindContours}
\title{FindContours}
\usage{
FindContours(dmat)
}
\arguments{
\item{XXXX}{}
}
\value{
XXXX
}
\description{
Wrapper for openCV function findContours
}
\details{
See documentation at: http://opencv.org/
}
\examples{
XXXX
}
\references{
XXXX
}

