% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/form.fit.R
\name{form.fit}
\alias{form.fit}
\title{form.fit}
\usage{
form.fit(dmat, deg, drop.val = "", num.x.pts = NULL, num.slices = NULL,
  x.phys.max = NULL, y.phys.max = NULL)
}
\arguments{
\item{dmat}{A matrix of points.}

\item{deg}{Degree of surface to fit. 1-XXXXXX.}

\item{drop.val}{Value of martix element in dmat to drop. Usually NaN or NA.}

\item{num.x.pts}{Number of points desired in the x-direction.}

\item{num.slices}{Number of points desired in the y-direction.}

\item{x.phys.max}{For fitting to physical length/width.}

\item{y.phys.max}{For fitting to physical length/width.}
}
\value{
Matrix of points laying on the fit surface.
}
\description{
Fit a 2D linear or polynomial form for a surface data set.
}
\details{
Handy to subtract a form
}
\examples{
XXXX
}
\references{
XXXX
}

