% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hole.mask.R
\name{hole.mask}
\alias{hole.mask}
\title{hole.mask}
\usage{
hole.mask(dmat, big.hole.tol.percent)
}
\arguments{
\item{XXXX}{}
}
\value{
XXXX
}
\description{
XXXX
}
\details{
Holes are 1. Measured points are 0.
}
\examples{
XXXX
}
\references{
XXXX
}

