% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mat2coords.R
\name{mat2coords}
\alias{mat2coords}
\title{mat2coords}
\usage{
mat2coords(dmat, x.phys.max = NULL, y.phys.max = NULL)
}
\arguments{
\item{dmat}{A matrix of points.}

\item{x.phys.max}{XXXX}

\item{y.phys.max}{XXXX}

\item{num.x.pts}{Number of points desired in the x-direction.}

\item{num.slices}{Number of points desired in the y-direction.}
}
\value{
XXXX
}
\description{
Convert a surface or image matrix to xyz coordinates
}
\details{
Output x y coordinates can be either indices or physical.
}
\examples{
XXXX
}
\references{
XXXX
}

