% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot3D.sip.R
\name{plot3D.sip}
\alias{plot3D.sip}
\title{plot3D.sip}
\usage{
plot3D.sip(dmat, num.x.pts = NULL, num.slices = NULL, aspect = c(1, 0.3,
  0.2), plot.type = "points")
}
\arguments{
\item{dmat}{A matrix of points.}

\item{num.x.pts}{Number of points desired in the x-direction.}

\item{num.slices}{Number of points desired in the y-direction.}

\item{aspect}{Aspect ratios}

\item{plot.type}{"points" or "surface"}
}
\value{
a 3D plot.
}
\description{
3D plot of a surface, image or points.
}
\details{
3D plot  of a surface image or points. Just send in XXXX. Some type of downsampling or decimation
is usually required so that the surface can be rendered fast. This function leverages openCV functionality
to do that. The The arguments num.x.pts (numcols) and num.slices (num.rows) allow the user to reduce the
surface down to the desited size. If the number of points is still potentially large, the function will prompt the
user if they want to continue.
}
\examples{
XXXX
}
\references{
XXXX
}

