% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rescale.z.R
\name{rescale.z}
\alias{rescale.z}
\title{rescale.z}
\usage{
\method{rescale}{z}(dmat, z.bitdepth, zero.maskQ = FALSE)
}
\arguments{
\item{dmat}{A matrix of data from R}

\item{z.bitdepth}{Desired bit depth. May be "8bit" or "16bit" for "unsigned" dynamic ranges of those bit depths.}

\item{zero.maskQ}{If TRUE, the smallest value of the dynamic range (0) is not used, with the intension that it is held in reserve for NaNs/NAs in the matrix.}
}
\value{
re-scaled dmat with matrix elements in the unsigned 8 or 16-bit range.
}
\description{
Put the measured z-valued into a 8 or 16-bit dynamic range.
}
\details{
Useful and sometimes necessary preprocessing step for analyzing surface data with openCV.
}
\examples{
XXXX
}
\references{
XXXX
}

